camera/requirements.txt
numpy>=1.26
Pillow>=10.0
piexif>=1.1.3
opencv-python-headless>=4.10
onnx>=1.17
onnxruntime>=1.20
tqdm>=4.66
